backend/requirements.txt
# FastAPI
fastapi>=0.100.0
uvicorn==0.22.0

# Functions Framework
functions-framework==3.*

# Core dependencies for the processing service
dependency-injector>=4.41.0
firebase-admin>=6.2.0
google-cloud-storage>=2.13.0
google-cloud-pubsub>=2.18.0
pydantic>=2.0.0

# PDF processing
PyPDF2>=3.0.0
pymupdf>=1.22.1

# OpenAI for content processing
openai>=1.3.0

# Utility packages
python-dotenv>=1.0.0
pydantic-settings>=2.0.0